CSN: CSN-hasim-alpha-45
Half of opc1c functions are supposed to be emulated.  My last check-in broke
the emulated instructions by claiming the emulated instructions have
destinations.  This sometimes breaks because the emulation code only updates
registers when they change value.  It should probably be fixed to force
updates on true destination registers, but for now simply claiming no
destinations in the decoder works, too.